feat: change selectFolderModal to a section inside repacksModal
