+++ src/main/events/hardware/get-disk-free-space.ts
@@ -1,10 +1,9 @@
 import checkDiskSpace from "check-disk-space";
 
 import { registerEvent } from "../register-event";
-import { getDownloadsPath } from "../helpers/get-downloads-path";
 
-const getDiskFreeSpace = async (_event: Electron.IpcMainInvokeEvent) =>
-  checkDiskSpace(await getDownloadsPath());
+const getDiskFreeSpace = async (_event: Electron.IpcMainInvokeEvent, path: string) =>
+  checkDiskSpace(path);
 
 registerEvent(getDiskFreeSpace, {
   name: "getDiskFreeSpace",
