+++ src/renderer/pages/game-details/game-details.tsx
@@ -51,7 +51,6 @@ export function GameDetails() {
   const { t, i18n } = useTranslation("game_details");
 
   const [showRepacksModal, setShowRepacksModal] = useState(false);
-  const [showSelectFolderModal, setShowSelectFolderModal] = useState(false);
 
   const randomGameObjectID = useRef<string | null>(null);
 
@@ -154,7 +153,6 @@ export function GameDetails() {
     ).then(() => {
       getGame();
       setShowRepacksModal(false);
-      setShowSelectFolderModal(false);
     });
   };
 
@@ -179,8 +177,6 @@ export function GameDetails() {
           visible={showRepacksModal}
           gameDetails={gameDetails}
           startDownload={handleStartDownload}
-          showSelectFolderModal={showSelectFolderModal}
-          setShowSelectFolderModal={setShowSelectFolderModal}
           onClose={() => setShowRepacksModal(false)}
         />
       )}
