+++ src/renderer/pages/game-details/repacks-modal.css.ts
@@ -16,3 +16,23 @@ export const repackButton = style({
   color: vars.color.bodyText,
   padding: `${SPACING_UNIT * 2}px`,
 });
+
+export const container = style({
+  width: "100%",
+  display: "flex",
+  flexDirection: "column",
+  gap: `${SPACING_UNIT * 2}px`,
+  marginBottom: SPACING_UNIT * 2,
+  paddingBottom: SPACING_UNIT * 2,
+  borderBottom: `solid 1px ${vars.color.borderColor}`,
+});
+
+export const downloadsPathField = style({
+  display: "flex",
+  gap: `${SPACING_UNIT * 2}px`,
+});
+
+export const hintText = style({
+  fontSize: 12,
+  color: vars.color.bodyText,
+});
